WebSockets/tests/submissions/Opera/interfaces/WebSocket/constants/001.html

<!doctype html>
<title>WebSockets: getting constants on constructor</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
test(function(){
  var ws = new WebSocket(SCHEME_AND_DOMAIN+'/');
  var constants = ['CONNECTING', 'OPEN', 'CLOSING', 'CLOSED'];
  for (var i = 0; i < constants.length; ++i) {
    assert_equals(WebSocket[constants[i]], i, 'WebSocket.'+constants[i]);
  }
});
end();
</script>
